+++ src/qz/printer/action/PrintPDF.java
@@ -16,8 +16,8 @@ import org.codehaus.jettison.json.JSONArray;
 import org.codehaus.jettison.json.JSONException;
 import org.codehaus.jettison.json.JSONObject;
 import qz.common.Constants;
-import qz.printer.BookBundle;
-import qz.printer.PDFWrapper;
+import qz.printer.action.entity.BookBundle;
+import qz.printer.action.entity.PDFWrapper;
 import qz.printer.PrintOptions;
 import qz.printer.PrintOutput;
 import qz.utils.ConnectionUtilities;
