+++ src/qz/printer/action/PrintRaw.java
@@ -24,10 +24,12 @@ import qz.common.ByteArrayBuilder;
 import qz.common.Constants;
 import qz.exception.NullCommandException;
 import qz.exception.NullPrintServiceException;
-import qz.printer.ImageWrapper;
-import qz.printer.LanguageType;
+import qz.printer.action.entity.ImageWrapper;
+import qz.printer.action.entity.LanguageType;
 import qz.printer.PrintOptions;
 import qz.printer.PrintOutput;
+import qz.printer.action.entity.WebApp;
+import qz.printer.action.entity.WebAppModel;
 import qz.printer.info.NativePrinter;
 import qz.printer.status.CupsUtils;
 import qz.utils.*;
