+++ src/qz/printer/action/entity/WebApp.java
@@ -1,4 +1,4 @@
-package qz.printer.action;
+package qz.printer.action.entity;
 
 import com.github.zafarkhaja.semver.Version;
 import com.sun.javafx.tk.TKPulseListener;
